requirements.txt
camelot-py
opencv-python
selenium
pytest
tqdm
click
dash
dash-daq
pandas